+++ package.json
@@ -13,8 +13,8 @@
     "format": "prettier \"**/*.+(ts|js|tsx|mdx|md)\"",
     "format.write": "yarn run format --write",
     "format.check": "yarn run format --check",
-    "generate-llms": "yarn run open-api-to-md && tsx scripts/generateLlmsTxt.ts && tsx scripts/generateApiReference.ts",
-    "generate-reference-md": "tsx scripts/generateApiReference.ts",
+    "generate-llms": "yarn run open-api-to-md && tsx scripts/generateLlmsTxt.ts && tsx scripts/generateApiMarkdown.ts",
+    "generate-reference-md": "tsx scripts/generateApiMarkdown.ts",
     "index-apis": "tsx scripts/indexApisForSearch.ts",
     "open-api-to-md": "bash scripts/openApiToMd.sh",
     "predev": "yarn generate-llms",
